Now with concurrency and progress bar

+++ pdf2odt/core.py
@@ -14,17 +14,15 @@ from pdf2odt.version import __versiondate__, __version__
 from officegenerator import ODT_Standard
 from odf.text import P
 from PIL import Image
-from concurrent.futures import ProcessPoolExecutor
 from multiprocessing import cpu_count
-from subprocess import check_output
-
+from subprocess import check_output, STDOUT
+from tqdm import tqdm
 
 try:
     t=gettext.translation('pdf2odt',pkg_resources.resource_filename("pdf2odt","locale"))
     _=t.gettext
 except:
     _=str
-
 
 def get_pdf_num_pages(filename):
     if platform.system()=="Windows":
@@ -34,18 +32,34 @@ def get_pdf_num_pages(filename):
 
     print("Converting pdf to images")
     try:
-        output=check_output(pdfinfo_command, shell=True)
-        print(output)
+        output=check_output(pdfinfo_command, shell=True).decode('UTF-8')
     except:
         pass
     
+    for line in output.split("\n"):
+        if line.find("Pages:")!=-1:
+            return int(line.split("Pages:")[1])
+    return 0
 
-
+def process_page(args, number, numpages):
+    zfill=str(number).zfill(len(str(numpages)))
+    pngfile="pdf2odt_temporal-{}.png".format(zfill)
+    
+    if platform.system()=="Windows":
+        pdftoppm_command='""pdftoppm.exe" -f {0} -l {0} -r {0} -png "{}" pdf2odt_temporal"'.format( args.resolution,  number,  args.pdf) #I add quotes to embrace all command too
+        tesseract_command='""tesseract.exe" {0} -l {1} {0}"'.format(pngfile, args.tesseract_language)#I add quotes to embrace all command too
+    else:
+        pdftoppm_command="pdftoppm -png -r {0} -f {1} -l {1} '{2}' pdf2odt_temporal".format( args.resolution, number, args.pdf)
+        tesseract_command="tesseract {0} -l {1} {0}".format(pngfile, args.tesseract_language)
+    check_output(pdftoppm_command, shell=True,  stderr=STDOUT)
+    if args.tesseract==True:
+        check_output(tesseract_command, shell=True,  stderr=STDOUT)
+    return number
 
 ## pdf2odt main script
 ## If arguments is None, launches with sys.argc parameters. Entry point is pdf2odt:main
 ## You can call with main(['--pretend']). It's equivalento to os.system('pdf2odt --pretend')
-## @param arguments is an array with parser arguments. For example: ['--max_files_to_store','9']. 
+## @param arguments is an array with parser arguments. 
 def main(arguments=None):
     start=datetime.datetime.now()
     parser=argparse.ArgumentParser(prog='pdf2odt', description=_('Converts a pdf to a LibreOffice Writer document with pages as images'), epilog=_("Developed by Mariano Muñoz 2019-{}".format(__versiondate__.year)), formatter_class=argparse.RawTextHelpFormatter)
@@ -54,44 +68,35 @@ def main(arguments=None):
     parser.add_argument('--tesseract_language', help=_("Language used in t1esseract command. Default is spa"), action="store", default="spa")
     parser.add_argument('--resolution', help=_("Sets DPI image resolution. Default is 300"), action="store", default="300")
     parser.add_argument('--tesseract', help=_("Uses tesseract ocr and insert result after image in ODT document"), action="store_true", default=False)
-    parser.add_argument('--output', help=_("Output odt file"), action="store")
+    parser.add_argument('output', help=_("Output odt file"), action="store")
 
     args=parser.parse_args(arguments)
 
     colorama_init(autoreset=True)
 
+    numpages=get_pdf_num_pages( args.pdf)
+    print(Style.BRIGHT +"Detected {} pages in {}".format(Fore.GREEN + str(numpages) + Fore.WHITE, args.pdf))
     # Sets locale to get integer format localized strings
     try:
         locale.setlocale(locale.LC_ALL, ".".join(locale.getlocale()))
     except:
         pass
+        
+    #Launching concurrent process
+    futures=[]
+    from concurrent.futures import ProcessPoolExecutor, as_completed
+    executor = ProcessPoolExecutor(max_workers=cpu_count()+1)
+    for number in range(numpages):
+        futures.append(executor.submit(process_page, args, number+1, numpages))
+    for f in tqdm(as_completed(futures), total=len(futures)):
+        pass
 
-    if platform.system()=="Windows":
-        pdftoppm_command='""pdftoppm.exe" -r {} -png "{}" pdf2odt_temporal"'.format( args.resolution,  args.pdf) #I add quotes to embrace all command too
-    else:
-        pdftoppm_command="pdftoppm -png -r {} '{}' pdf2odt_temporal".format( args.resolution, args.pdf)
-
-    print("Converting pdf to images")
-    os.system(pdftoppm_command)
-
+    #Generating ODT
     doc=ODT_Standard(args.output)
     pdf=os.path.basename(args.pdf)
     odt=os.path.basename(args.output)
     doc.setMetadata(_("Converting PDF to ODT"),  _("Converting {} to {} using odt2pdf-{}").format(pdf, odt, __version__), "odt2pdf")
     
-    
-    get_pdf_num_pages( args.pdf)
-    
-    
-    futures=[]
-    with ProcessPoolExecutor(max_workers=cpu_count()+1) as executor:
-        for filename in sorted(glob.glob("pdf2odt_temporal*.png")):
-            if platform.system()=="Windows":
-                tesseract_command='""tesseract.exe" {1} -l {2} {1}"'.format(filename, args.tesseract_language)#I add quotes to embrace all command too
-            else:
-                tesseract_command="tesseract {1} -l {2} {1}".format(filename, args.tesseract_language)
-            futures.append(executor.submit(os.system, tesseract_command))    
-
     for filename in sorted(glob.glob("pdf2odt_temporal*.png")):
         img = Image.open(filename)
         x,y=img.size
@@ -102,16 +107,18 @@ def main(arguments=None):
         p = P(stylename="Illustration")
         p.addElement(doc.image(filename, cmx,cmy))
         doc.insertInCursor(p, after=True)
-        
-        for line in open(filename +".txt", "r").readlines():
-            p=P(stylename="Standard")
-            p.addText(line)
-            doc.insertInCursor(p, after=True)
+        if args.tesseract==True:
+            for line in open(filename +".txt", "r").readlines():
+                p=P(stylename="Standard")
+                p.addText(line)
+                doc.insertInCursor(p, after=True)
         
     doc.save()
 
+    #Removing temporal files
     for filename in glob.glob("pdf2odt_temporal*.png"):
         os.remove(filename)
-        os.remove(filename+".txt")
+        if args.tesseract==True:
+            os.remove(filename+".txt")
         
     print(Style.BRIGHT + _("ODT generation took {}").format(Fore.GREEN + str(datetime.datetime.now()-start)))
